\name{CenMixReg-package}
\alias{CenMixReg-package}
\alias{CenMixReg}
\docType{package}
\title{
Modeling Censored Data Using Mixture Regression
}
\description{
This package contains a principal function that performs to estimate estimate the parameters of a regression model considering an error that follows a finite mixture of Normal or Student-t distributions, using an analytically simple and efficient EM-type algorithm for iteratively computing maximum likelihood estimates of the parameters.
}
\details{
\tabular{ll}{
Package: \tab CensMixReg\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2015-08-26\cr
License: \tab GPL (>=2)\cr
}
}

\author{Luis Benites Sanchez \email{lbenitesanchez@gmail.com}, Victor Hugo Lachos \email{hlachos@ime.unicamp.br}
}

\references{
Karlsson, M. & Laitila, T. (2014). Finite mixture modeling of censored regression models. Statistical papers, 55(3), 627-642.

Massuia, M. B., Cabral, C. R. B., Matos, L. A. & Lachos, V. H. (2014). Influence diagnostics for student-t censored linear regression models. Statistics, (ahead-of-print), 1-21.

Arellano-Valle, R., Castro, L., Gonzalez-Farias, G. & Munoz-Gajardo, K. (2012). Student-t censored regression model: properties and inference. Statistical Methods & Applications, 21, 453-473.

Garay, A. M., Lachos, V. H., Bolfarine, H. & Cabral, C. R. (2015). Linear censored regression models with scale mixtures of normal distributions. Statistical Papers, pages 1-32.
}

\keyword{package}

\seealso{
\code{\link{CensMixReg}}
}

\examples{
#See examples for the CensMixReg function linked above.

}
